requirements.txt
streamlit==1.31.0
python-dotenv==1.0.0
pillow==10.3.0
#torch==2.3.0   #pip3 install --pre torch torchvision torchaudio --index-url https://download.pytorch.org/whl/nightly/cu124
huggingface-hub==0.23.0
PyMuPDF==1.22.5
docarray~=0.37.0
openai~=0.27.8
llama-cpp-python~=0.2.76
tiktoken==0.4.0
langchain_text_splitters
bitsandbytes   #pip install bitsandbytes-cuda113
langchain_community==0.2.1
langchain==0.0.278  #0.1.12
gpt4all

#auto-gptq==0.3.2
#pdfminer.six
#sentence_transformers0=2.7.0

#CUDA_LAUNCH_BLOCKING=1.
#`TORCH_USE_CUDA_DSA`
